Updating the service layer
Some unused files are removed from the service layer. The operation of updating a product is updated in a way that it will not allow a user to modify a product if he already has another product with the same name. The operation of deleting a product is updated in a way that reports of the product will be deleted when the product is deleted (previously the delete product operation was throwing errors if the product had any reports). Another change is that the app will not allow a user to update or delete a product if the product has non resolved reports.

+++ TradingApp.Services.Core/ProductOperationsService.cs
@@ -71,8 +71,19 @@ namespace TradingApp.Services.Core
                 return new Result(errorCode: ProductErrorCodes.ProductInvalidCreator);
             }
 
-            int productActiveSellOrdersCount = await _productService.GetProductActiveSellOrdersCountAsync(productId: id);
+            bool doesCreatorHaveOtherProductsWithTheSameName = await _context.Products.AsNoTracking().AnyAsync(p => p.Name == name && p.Id != id && p.CreatorId == creatorId);
+            if(doesCreatorHaveOtherProductsWithTheSameName == true) 
+            {
+                return new Result(errorCode: ProductErrorCodes.ProductWithSameNameAlreadyExists);
+            }
 
+            bool doesProductHaveNonResolvedReports = await _context.ProductReports.AsNoTracking().AnyAsync(pr => pr.ReportedProductId == id && pr.Status != ProductReportStatus.resolved);
+            if (doesProductHaveNonResolvedReports == true)
+            {
+                return new Result(errorCode: ProductErrorCodes.ProductHasNonResolvedReports);
+            }
+
+            int productActiveSellOrdersCount = await _productService.GetProductActiveSellOrdersCountAsync(productId: id);
             if (productActiveSellOrdersCount > 0)
             {
                 return new Result(errorCode: ProductErrorCodes.ProductHasActiveSaleOrders);
@@ -103,8 +114,16 @@ namespace TradingApp.Services.Core
                 return new Result(errorCode: ProductErrorCodes.ProductInvalidCreator);
             }
 
+            bool doesProductHaveNonResolvedReports = await _context.ProductReports.AsNoTracking().AnyAsync(pr => pr.ReportedProductId == id && pr.Status != ProductReportStatus.resolved);
+            if(doesProductHaveNonResolvedReports == true)
+            {
+                return new Result(errorCode: ProductErrorCodes.ProductHasNonResolvedReports);
+            }
+
+
             await DeleteSellOrdersOfProductAsync(id);
             await DeleteSellOrderSuggestionsOfProductAsync(id);
+            await DeleteReportsOfProductAsync(id);
             await SetFkToNullForCompletedOrdersOfProductAsync(id);
 
             _context.Products.Attach(product);
@@ -134,6 +153,7 @@ namespace TradingApp.Services.Core
             return new Result();
         }
 
+
         private async Task DeleteSellOrdersOfProductAsync(Guid productId)
         {
             IEnumerable<SellOrder> sellOrders = await _context
@@ -155,6 +175,16 @@ namespace TradingApp.Services.Core
             _context.SellOrderSuggestions.RemoveRange(sellOrderSuggestions);
         }
 
+        private async Task DeleteReportsOfProductAsync(Guid productId)
+        {
+            IEnumerable<ProductReport> productReports = await _context
+                .ProductReports
+                .Where(pr => pr.ReportedProductId == productId)
+                .ToListAsync();
+
+            _context.ProductReports.RemoveRange(productReports);
+        }
+
         private async Task SetFkToNullForCompletedOrdersOfProductAsync(Guid productId)
         {
             IEnumerable<CompletedOrder> completedOrders = await _context
@@ -168,5 +198,7 @@ namespace TradingApp.Services.Core
                 completedOrder.Product = null;
             }
         }
+
+        
     }
 }
